New information added.
I have added a slide to represent benefits of features. A couple of rubber duck
slides are now at the beginning for the new intro. An "implementation detail"
star has been added for the features slide. correct attribution has been added
for images (except for the ones I haven't found yet the name of the author)